Move Python scripts to Frontiers version control
This is a major "dump": previously the scripts used for data processing
and analysis in this project were kept in a separate project.
I'm moving them here now, and this commit is essentially a "here you go"
dump of the entire suite of Python scripts.
Worth mentioning is make_tex_tables.py, which I assembled today to
automate making LaTeX tables of parameters.