Add function to autogenerate SPM subplot figure
Motivation: Frontiers requires that sub-figures be submitted as a single
figure (instead of e.g. combining individual figures with the `subcaption`
package in LaTeX).
This commit creates a function to autogenerate an 8-axis subfigure, of
SPM test results of which would be prohibitively time-consuming to
repeatedly put together by hand.