Snapshot of project-wide pre-cleanup code state
Context: The current source code is a jumble scripts originally from
different directories. I have now moved general TMG parameter
computation to a standalone Python package, and in this project intend
to keep only code that is specific to the analysis performed for the
Frontiers article.
This will probably trim down the code considerably.